python-examples/requirements.txt
Appium-Python-Client==0.51
coverage==4.5.4
nose2==0.9.1
python-dateutil==2.8.0
selenium==3.141.0
six==1.12.0
text-unidecode==1.2
urllib3==1.25.3
